requirements.txt
PyYAML==6.0.2
SQLAlchemy==2.0.36
confluent_kafka==2.6.1
pandas==2.2.3
psycopg2-binary==2.9.10
